pyproject.toml
[project]
name = "langgraph-study"
version = "0.1.0"
description = ""
authors = [
    {name = "Rakshak1344",email = "dev.rakshith14@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
packages = [
    { include = "reactagent", from = "src" },
    { include = "reflection_agent", from = "src" },
]
dependencies = [
    "langgraph (>=0.3.25,<0.4.0)",
    "isort (>=6.0.1,<7.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "langchain-openai (>=0.3.12,<0.4.0)",
    "grandalf (>=0.8,<0.9)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "tavily-python (>=0.5.4,<0.6.0)",
    "black (>=25.1.0,<26.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
